Move types and type guards to module
